feat: customer_segment initial submission
Customer_segment

The first submission of the assignment. The problem was to classify the
wholesale customer classification of wholesale and practice customer
clustering through unsupervised learning based on the information
obtained here. We analyzed the data using PCA, k-Means and Gaussian
Mixture Model and compared the customer classification of the original
data.