fix: 2nd submission
This is the second submission reflecting the first review.

Reflecting the use of seaborn to represent the sample data.

We used the results from the principal component analysis to reflect the idea of describing the influence of feature points using terms such as corelation. (With reference to skewed data)

When outlier is excluded, except for all cases, 10% of the total data will be removed. Reflecting the opinion that this is not good, we revised the implementation by using collections.Counter to judge the duplicated data more than twice as outlier and remove it. .

And that the interpretation of the PCA results is necessary to be supplemented.

It reflects the requirement to describe how to use unsupervised learning
in A / B testing.