fix: Modified PCA code
We have modified the PCA code so that we can analyze all 6 features. And
we analyzed different customer by individual factors up to 4 dimensions.

There is an additional code that reflects the comment of the reviewer in
the code and newly reflects the data analysis api contents which it did
not know before.

Data.corr ()