fix : Update the Driving Agent State
For your second implementation, navigate to the 'build_state()' agent
function. With the justification you've provided in Question 4, you will
now set the 'state' variable to a tuple of all the features necessary
for Q-Learning. Confirm your driving agent is updating its state by
running the agent file and simulation briefly and note whether the state
is displaying. If the visual simulation is used, confirm that the
updated state corresponds with what is seen in the simulation.